refactor: 특정클래스에서만 사용하는 클래스  inner class로 변경
