+++ src/test/java/stringcalculator/operator/CustomOperator.java
@@ -1,13 +1,12 @@
 package stringcalculator.operator;
 
 import stringcalculator.number.Number;
-import stringcalculator.operator.pattern.NumberPattern;
 import stringcalculator.number.ZeroNumber;
-import stringcalculator.operator.pattern.PrefixPattern;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public class CustomOperator implements Operator {
 
@@ -18,7 +17,7 @@ public class CustomOperator implements Operator {
     }
 
     public static CustomOperator from(String input) {
-        return new CustomOperator(makeNumbers(new PrefixPattern(input).removeUnnecessary()));
+        return new CustomOperator(makeNumbers(new PrefixPattern(input).value()));
     }
 
     private static List<Number> makeNumbers(String input) {
@@ -50,6 +49,46 @@ public class CustomOperator implements Operator {
     @Override
     public int hashCode() {
         return numbers != null ? numbers.hashCode() : 0;
+    }
+
+    private static class PrefixPattern {
+
+        private final String origin;
+        private final Pattern pattern;
+
+        public PrefixPattern(String origin, Pattern pattern) {
+            this.origin = origin;
+            this.pattern = pattern;
+        }
+
+        public PrefixPattern(String origin) {
+            this(origin, Pattern.compile("//(.)\\\\n"));
+        }
+
+        public String value() {
+            return pattern.matcher(origin).replaceAll("");
+        }
+
+    }
+
+    private static class NumberPattern {
+
+        private final String origin;
+        private final Pattern pattern;
+
+        public NumberPattern(String origin, Pattern pattern) {
+            this.origin = origin;
+            this.pattern = pattern;
+        }
+
+        public NumberPattern(String origin) {
+            this(origin, Pattern.compile("\\d+"));
+        }
+
+        public Matcher getMatcher() {
+            return pattern.matcher(origin);
+        }
+
     }
 
 }
